// src/typescripts/types/common.types.ts
export type LoginMode = 'mobile' | 'email';

export type TDayjsFormat =
  | 'DD MMM YYYY'
  | 'YYYY-MM-DD'
  | 'DD/MM/YYYY'
  | 'MMM D, YYYY'
  | 'dddd, MMM D'
  | 'YYYY/MM/DD'
  | 'D MMMM YYYY'
  | 'MMM YYYY'
  | 'DDD MM YYYY'
  | 'ddd, DD MMM'
  | 'D MMMM, dddd'
  | 'MMM DD, YYYY  hh : mm A'
  | 'ddd MMM'
  | 'DD-MM-YYYY'
  | 'MMMM DD, YYYY';